requirements.txt
# Miel Monteur Build Requirements
pillow>=10.0.0
pycdlib>=1.14.0
pydub>=0.25.0
PyTexturePacker>=1.2.0
requests>=2.28.0
GitPython>=3.1.0
bitstring>=4.0.0
libsass
